requirements.txt
colorama
demjson3
faiss-cpu
geopy
numpy
openai
pybars3
pymupdf
python-dotenv
pytz
questionary
rich
tiktoken
timezonefinder
tqdm
typer
llama_index
